Tidy up hoverbelow code
