+++ src/app/static/src/app/components/figures/graphs/plotlyGraph.vue
@@ -1,20 +1,22 @@
 <template>
-    <plotly :data="dataSeries" :layout="layout" :display-mode-bar="true"></plotly>
+    <div :class="{hoverbelow: hoverBelow}">
+        <plotly :data="dataSeries" :layout="layout" :display-mode-bar="true"></plotly>
+    </div>
 </template>
 <script lang="ts">
     import {Plotly} from "vue-plotly"
-    import {defineComponent, onBeforeUnmount, onMounted} from "@vue/composition-api";
+    import {defineComponent} from "@vue/composition-api";
     import {FilteringProps} from "../filteredData";
     import {useLongFormatData} from "./longFormatDataSeries";
     import {useWideFormatData} from "./wideFormatDataSeries";
+    import {setupHoverBelowObserver} from "./hoverBelow";
     import {Dictionary} from "vue-router/types/router";
     import {LongFormatMetadata, SeriesDefinition, WideFormatMetadata} from "../../../generated";
-    import $ from 'jquery';
 
     interface Props extends FilteringProps {
         series: SeriesDefinition[]
         metadata: WideFormatMetadata | LongFormatMetadata
-        layout: Dictionary<string>
+        layout: Dictionary<any>
     }
 
     export default defineComponent({
@@ -23,56 +25,11 @@
         },
         props: {settings: Object, data: Array, series: Array, metadata: Object, layout: Object},
         setup(props: Props) {
-            //TODO: Only do this if configured in layout
-            //TODO: put file styles in class and apply only if configured
-            //TODO: sort out destroyed
-            //TODO: move out to another file?
-
-            onMounted(() => {
-                const target = $( ".plotly-horizontal-errors" )[0];
-
-                if (target) {
-                    const config = {
-                        attributes: true,
-                        attributeFilter: ["y", "d"],
-                        childList: true,
-                        characterData: false,
-                        subtree: true
-                    };
-
-                    const observer = new MutationObserver(function (mutations) {
-                        mutations.forEach(function (mutation) {
-                            const text = $([mutation.target]).filter(".hoverlayer .hovertext text");
-                            if (text.length && text.attr("y") !== "0") {
-                                 text.attr("y", "0");
-                            }
-
-                            const path = $([mutation.target]).filter(".hoverlayer .hovertext path");
-                            if (path.length) {
-                                const w = path.siblings("rect").attr("x") || 0;
-                                const h = path.siblings("rect").height() || 0;
-                                const newPath = "M4,-" + h/2 + //path is clockwise, starting at top left
-                                    "v" + h  + //vertical line to bottom
-                                    "H" + w + //horizontal line to right
-                                    "v-" + h + //vertical line to top
-                                    "h-" + (+w-20) + //horizontal line left to start of 'up' indicator
-                                    "l-6,-6" +  //diagonal line to top of 'up' indicator
-                                    "l-6,6" + //diagonal line to bottom of 'up' indicator
-                                    "Z"; //close path
-                                if (path.attr("d") !== newPath) {
-                                    path.attr("d", newPath)
-                                }
-                            }
-                        });
-                    });
-
-                    observer.observe(target, config);
-                }
-            });
-
-            onBeforeUnmount(() => {
-
-            });
+            let hoverBelow = false;
+            if (props.layout.mint_custom && props.layout.mint_custom.hover_below) {
+                hoverBelow = true;
+                setupHoverBelowObserver("hoverbelow");
+            }
 
             let dataSeries = null;
             if (props.metadata.format == "long") {
@@ -82,7 +39,8 @@
             }
 
             return {
-                dataSeries
+                dataSeries,
+                hoverBelow
             }
         }
     })
